Added support for JSON path
An example to use it would be :
_$map :
  x : $.name
  y : $.age
  c : $.x[7]

mapOutput : derived.nameObject | jsonpath(derived._$map)

A few things are added on top of jsonpath query result:
1. If the query results an empty array, then it is converted to empty string.
2. If the query results an array with only one element, then only the element is returned and not the array.
3. If the query results an array with many elements, then its returned as is.

<value> | jsonpath(pathString | subTransformBlock)

The jsonpath can return a single path value or a map of values based on what is passed to it
as path argument. The path argument can be a simple jsonpath expression or a subTransform block
which contains a map of keys and jsonpaths as its values.